feat(EMS-3119-3122): no pdf - export contract - check your answers - agent charges
