+++ src/ui/server/helpers/summary-lists/export-contract/agent-fields/index.test.ts
@@ -1,4 +1,4 @@
-import agentFields, { agentDetailsFields, agentServiceFields, agentServiceChargeFields } from '.';
+import agentFields, { agentDetailsFields, agentServiceFields } from '.';
 import { FORM_TITLES } from '../../../../content-strings';
 import { EXPORT_CONTRACT_FIELDS as FIELDS } from '../../../../content-strings/fields/insurance';
 import FIELD_IDS from '../../../../constants/field-ids/insurance/export-contract';
@@ -6,10 +6,10 @@ import { EXPORT_CONTRACT as EXPORT_CONTRACT_ROUTES } from '../../../../constants
 import fieldGroupItem from '../../generate-field-group-item';
 import getFieldById from '../../../get-field-by-id';
 import mapYesNoField from '../../../mappings/map-yes-no-field';
-import mapPercentage from '../../../map-percentage';
 import getCountryByIsoCode from '../../../get-country-by-iso-code';
 import generateChangeLink from '../../../generate-change-link';
 import replaceNewLineWithLineBreak from '../../../replace-new-line-with-line-break';
+import agentChargesFields from './agent-charges';
 import { mockApplication, mockCountries, referenceNumber } from '../../../../test-mocks';
 
 const {
@@ -19,35 +19,17 @@ const {
 const {
   USING_AGENT,
   AGENT_DETAILS: { NAME, FULL_ADDRESS, COUNTRY_CODE },
-  AGENT_SERVICE: { SERVICE_DESCRIPTION, IS_CHARGING },
-  AGENT_CHARGES: { FIXED_SUM_AMOUNT, CHARGE_PERCENTAGE, PAYABLE_COUNTRY_CODE },
+  AGENT_SERVICE: { SERVICE_DESCRIPTION },
 } = FIELD_IDS;
 
-const {
-  AGENT_CHANGE,
-  AGENT_CHECK_AND_CHANGE,
-  AGENT_DETAILS_CHANGE,
-  AGENT_DETAILS_CHECK_AND_CHANGE,
-  AGENT_SERVICE_CHANGE,
-  AGENT_SERVICE_CHECK_AND_CHANGE,
-  AGENT_CHARGES_CHANGE,
-  AGENT_CHARGES_CHECK_AND_CHANGE,
-} = EXPORT_CONTRACT_ROUTES;
+const { AGENT_CHANGE, AGENT_CHECK_AND_CHANGE } = EXPORT_CONTRACT_ROUTES;
+const { AGENT_DETAILS_CHANGE, AGENT_DETAILS_CHECK_AND_CHANGE } = EXPORT_CONTRACT_ROUTES;
+const { AGENT_SERVICE_CHANGE, AGENT_SERVICE_CHECK_AND_CHANGE } = EXPORT_CONTRACT_ROUTES;
 
 describe('server/helpers/summary-lists/export-contract/agent-fields', () => {
   const mockAnswersUsingAgentTrue = {
     ...mockApplication.exportContract.agent,
     [USING_AGENT]: true,
-  };
-
-  const mockAnswersUsingAgentTrueAgentChargingTrue = {
-    ...mockAnswersUsingAgentTrue.service,
-    [IS_CHARGING]: true,
-  };
-
-  const mockAnswersUsingAgentTrueAgentChargingFalse = {
-    ...mockAnswersUsingAgentTrue.service,
-    [IS_CHARGING]: false,
   };
 
   const mockAnswersUsingAgentFalse = {
@@ -112,119 +94,6 @@ describe('server/helpers/summary-lists/export-contract/agent-fields', () => {
     });
   });
 
-  describe('agentServiceChargeFields', () => {
-    describe(`when ${IS_CHARGING} is false`, () => {
-      it(`should return ${IS_CHARGING} field and value`, () => {
-        const result = agentServiceChargeFields(mockAnswersUsingAgentTrueAgentChargingFalse, referenceNumber, mockCountries, checkAndChange);
-
-        const expected = [
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_SERVICE, IS_CHARGING),
-              data: mockAnswersUsingAgentTrueAgentChargingFalse,
-              href: generateChangeLink(AGENT_SERVICE_CHANGE, AGENT_SERVICE_CHECK_AND_CHANGE, `#${IS_CHARGING}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            mapYesNoField(mockAnswersUsingAgentTrueAgentChargingFalse[IS_CHARGING]),
-          ),
-        ];
-
-        expect(result).toEqual(expected);
-      });
-    });
-
-    describe(`when ${IS_CHARGING} is true and ${FIXED_SUM_AMOUNT} is provided`, () => {
-      const mockAnswers = {
-        ...mockAnswersUsingAgentTrueAgentChargingTrue,
-        charge: {
-          ...mockAnswersUsingAgentTrueAgentChargingTrue.charge,
-          [CHARGE_PERCENTAGE]: null,
-          [FIXED_SUM_AMOUNT]: 10,
-        },
-      };
-
-      it('should return all agent service charge fields and values', () => {
-        const result = agentServiceChargeFields(mockAnswers, referenceNumber, mockCountries, checkAndChange);
-
-        const expected = [
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_SERVICE, IS_CHARGING),
-              data: mockAnswers,
-              href: generateChangeLink(AGENT_SERVICE_CHANGE, AGENT_SERVICE_CHECK_AND_CHANGE, `#${IS_CHARGING}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            mapYesNoField(mockAnswers[IS_CHARGING]),
-          ),
-          fieldGroupItem({
-            field: getFieldById(FIELDS.AGENT_CHARGES, FIXED_SUM_AMOUNT),
-            data: mockAnswers.charge,
-            href: generateChangeLink(AGENT_CHARGES_CHANGE, AGENT_CHARGES_CHECK_AND_CHANGE, `#${FIXED_SUM_AMOUNT}-label`, referenceNumber, checkAndChange),
-            renderChangeLink: true,
-          }),
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_CHARGES, PAYABLE_COUNTRY_CODE),
-              data: mockAnswers.charge,
-              href: generateChangeLink(AGENT_CHARGES_CHANGE, AGENT_CHARGES_CHECK_AND_CHANGE, `#${PAYABLE_COUNTRY_CODE}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            getCountryByIsoCode(mockCountries, mockAnswers.charge[PAYABLE_COUNTRY_CODE]).name,
-          ),
-        ];
-
-        expect(result).toEqual(expected);
-      });
-    });
-
-    describe(`when ${IS_CHARGING} is true and ${CHARGE_PERCENTAGE} is provided`, () => {
-      const mockAnswers = {
-        ...mockAnswersUsingAgentTrueAgentChargingTrue,
-        charge: {
-          ...mockAnswersUsingAgentTrueAgentChargingTrue.charge,
-          [FIXED_SUM_AMOUNT]: null,
-          [CHARGE_PERCENTAGE]: 10,
-        },
-      };
-
-      it('should return all agent service charge fields and values', () => {
-        const result = agentServiceChargeFields(mockAnswers, referenceNumber, mockCountries, checkAndChange);
-
-        const expected = [
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_SERVICE, IS_CHARGING),
-              data: mockAnswers,
-              href: generateChangeLink(AGENT_SERVICE_CHANGE, AGENT_SERVICE_CHECK_AND_CHANGE, `#${IS_CHARGING}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            mapYesNoField(mockAnswers[IS_CHARGING]),
-          ),
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_CHARGES, CHARGE_PERCENTAGE),
-              data: mockAnswers.charge,
-              href: generateChangeLink(AGENT_CHARGES_CHANGE, AGENT_CHARGES_CHECK_AND_CHANGE, `#${CHARGE_PERCENTAGE}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            mapPercentage(mockAnswers.charge[CHARGE_PERCENTAGE]),
-          ),
-          fieldGroupItem(
-            {
-              field: getFieldById(FIELDS.AGENT_CHARGES, PAYABLE_COUNTRY_CODE),
-              data: mockAnswers.charge,
-              href: generateChangeLink(AGENT_CHARGES_CHANGE, AGENT_CHARGES_CHECK_AND_CHANGE, `#${PAYABLE_COUNTRY_CODE}-label`, referenceNumber, checkAndChange),
-              renderChangeLink: true,
-            },
-            getCountryByIsoCode(mockCountries, mockAnswers.charge[PAYABLE_COUNTRY_CODE]).name,
-          ),
-        ];
-
-        expect(result).toEqual(expected);
-      });
-    });
-  });
-
   describe('agentFields', () => {
     describe(`when ${USING_AGENT} is false`, () => {
       it(`should return one ${USING_AGENT} field and value`, () => {
@@ -267,7 +136,7 @@ describe('server/helpers/summary-lists/export-contract/agent-fields', () => {
             ),
             ...agentDetailsFields(mockAnswersUsingAgentTrue, referenceNumber, mockCountries, checkAndChange),
             ...agentServiceFields(mockAnswersUsingAgentTrue.service, referenceNumber, checkAndChange),
-            ...agentServiceChargeFields(mockAnswersUsingAgentTrue.service, referenceNumber, mockCountries, checkAndChange),
+            ...agentChargesFields(mockAnswersUsingAgentTrue.service, referenceNumber, mockCountries, checkAndChange),
           ],
         };
